Extracted Web3Mock to new mocks directory
